Droid spawning after explosion comment
There is a bug where more droids spawn in becuase multiple lasers may hit the same droid bounding circle. which causes explosion to be called multiple times on the same droid before it and its bounding circle is removed. Currently the explode method of the AbstractDroid causes a droid to be spawned which will need to be changed after prototype is submitted.